refactor settings repository
